Fixed a ui bug and made special classes to remove code from the main window
