+++ MainWindow.xaml
@@ -18,6 +18,7 @@
         <Button x:Name="Change" Content="Change" Visibility="Visible" FontSize="16" Margin="0,0,73,69" IsEnabled="False" Click="Change_Click" Grid.Row="1" HorizontalAlignment="Right" Width="177" Height="60" VerticalAlignment="Bottom"/>
         <Button x:Name="Stop" Content="Stop" Visibility="Collapsed" FontSize="16" Margin="0,0,73,69" IsEnabled="True" Click="Stop_Click" Grid.Row="1" HorizontalAlignment="Right" Width="177" Height="60" VerticalAlignment="Bottom"/>
         <Image x:Name="Window2" Source="Images/background2.jpg" VerticalAlignment="Center" HorizontalAlignment="Center" Height="Auto" Width="Auto" Margin="74,50,74,160" Grid.RowSpan="2"/>
+        <Image x:Name="Window3" Source="Images/background2.jpg" VerticalAlignment="Center" HorizontalAlignment="Center" Height="Auto" Width="Auto" Margin="74,50,74,160" Grid.RowSpan="2"/>
         <Label x:Name="NoWallpaper" Content="No Wallpaper Selected" HorizontalAlignment="Center" VerticalAlignment="Center" Margin="270,37,270,279" HorizontalContentAlignment="Center" VerticalContentAlignment="Center" Height="60" FontSize="30" Grid.Row="1"/>
         <Label x:Name="Warning" HorizontalAlignment="Center" Margin="0,0,0,135" Grid.Row="1" VerticalAlignment="Bottom"/>
         <Button x:Name="Picture" Content="Picture" IsEnabled="False" HorizontalAlignment="Left" VerticalAlignment="Top" Click="Picture_Click" Height="40" Width="80"/>
